api/requirements.txt
Flask==2.1.1
feedparser==6.0.8
python-dateutil==2.8.2